Moving all migration functions out of `cmd_export.py` and into `aiida.cmdline.utils.migration`.
Following files have been created:
- `__init__.py`: Contains `migrate_recursive` and allows for import from `aiida.cmdline.utils.migration`
- `utils.py`: Contains `verify_metadata_version` and `update_metadata`
- `v01_to_v02.py`: Contains function to migrate from export version v0.1 to v0.2
- `v02_to_v03.py`: Contains function to migrate from export version v0.2 to v0.3
- `v03_to_v04.py`: Contains functions to migrate from export version v0.3 to v0.4

Moving and renaming new test-file `test_export_file_migrations.py` from `aiida.backends.tests`
to `aiida.backends.tests.cmdline.utils.migrations`, renaming it to `test_v03_to_v04.py`.

See previous commits for actual authors of the code.
