+++ aiida/backends/tests/__init__.py
@@ -70,6 +70,7 @@ DB_TEST_LIST = {
         'cmdline.params.types.path': ['aiida.backends.tests.cmdline.params.types.test_path'],
         'cmdline.params.types.plugin': ['aiida.backends.tests.cmdline.params.types.test_plugin'],
         'cmdline.utils.common': ['aiida.backends.tests.cmdline.utils.test_common'],
+        'cmdline.utils.migration.v03_to_v04': ['aiida.backends.tests.cmdline.utils.migration.test_v03_to_v04'],
         'common.archive': ['aiida.backends.tests.common.test_archive'],
         'common.extendeddicts': ['aiida.backends.tests.common.test_extendeddicts'],
         'common.folders': ['aiida.backends.tests.common.test_folders'],
@@ -103,7 +104,6 @@ DB_TEST_LIST = {
         'engine.work_chain': ['aiida.backends.tests.engine.test_work_chain'],
         'engine.workfunctions': ['aiida.backends.tests.engine.test_workfunctions'],
         'export_and_import': ['aiida.backends.tests.test_export_and_import'],
-        'export_file_migrations': ['aiida.backends.tests.test_export_file_migrations'],
         'generic': ['aiida.backends.tests.test_generic'],
         'manage.configuration.config.': ['aiida.backends.tests.manage.configuration.test_config'],
         'manage.configuration.migrations.': ['aiida.backends.tests.manage.configuration.migrations.test_migrations'],
